fix(analytics): rename Google Analytics provider files to get around Ad blockers
