+++ src/providers/angulartics2-ga.metadata.json
