+++ src/providers/angulartics2-ga.ts
